refactor(gateway): extract chat dispatch errors
